Improve editing character names
So far, we did not remember the composition and also did not support selection.
Depending on the keyboard used, this could also lead to infinite loops when
the keyboard tries to reach a state that was not reachable.

We now support these states by remembering all 4 positions
(start selection, end selection and optionally start composition and end composition)